requirements.txt
mcp[cli]==1.3.0
requests>=2.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0
starlette>=0.27.0
uvicorn>=0.22.0
httpx>=0.24.0
PyYAML>=6.0